feat(Composite): Moving files between folders
